polar Plot finished, Theta now < 90, some bugs removed, Energy spectrum is divided by E

+++ src/main/java/ru/mauveferret/Distributions/Distribution.java
@@ -36,7 +36,7 @@ public abstract class Distribution {
     }
 
     public void check(){};
-    abstract int[] getSpectrum();
+    abstract double[] getSpectrum();
     public abstract boolean logDistribution();
     public abstract boolean visualize();
